backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.0
pydantic>=2.9.0
python-multipart>=0.0.12
httpx>=0.27.0
google-generativeai>=0.3.0
Pillow>=10.0.0
python-dotenv>=1.0.0
